+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added support for Baked Indirect mixed lighting.
 - You can now use Light Probes for occlusion. This means that baked lights can now occlude dynamic objects.
+- (WIP) Added an experimental 2D renderer that implements a 2D lighting system.
+- (WIP) Added a Light2D component that works with the 2D renderer to add lighting effects to 2D sprites.
 
 ### Fixed
 - Fixed a project import issue in the LWRP template.