+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.11.0] - 2019-XX-XX
+
 ## [5.10.0] - 2019-03-19
 ### Added
 - Added RenderObjects. You can add RenderObjects to a Renderer to perform custom rendering.